Moved lock utility into threading namespace
